requirements.txt
python-jenkins==1.7.0
jenkinsapi
requests==2.25.0
paramiko==2.3.3
PyYAML==6.0
cryptography==38.0.4